requirements-dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    make up
#
aiocontextvars==0.2.2
aiofiles==0.4.0
aiomysql==0.0.20
aiosqlite==0.10.0
alabaster==0.7.12         # via sphinx
aniso8601==6.0.0
apispec==2.0.2
apistar==0.7.2
appdirs==1.4.3            # via black
asn1crypto==0.24.0        # via cryptography
astroid==2.2.5            # via pylint
asyncpg==0.18.3
asynctest==0.13.0
atomicwrites==1.3.0       # via pytest
attrs==19.1.0
autopep8==1.4.4
babel==2.7.0              # via sphinx
bandit==1.6.2
black==19.3b0 ; python_version >= "3.6"
bleach==3.1.0             # via readme-renderer
blinker==1.4              # via quart
cattrs==0.9.0
certifi==2019.6.16
cffi==1.12.3              # via cryptography
chardet==3.0.4
ciso8601==2.1.1
click==7.0
colorama==0.4.1           # via green
contextvars==2.4
coverage==4.5.3           # via coveralls, green, nose2
coveralls==1.8.1
cryptography==2.7         # via pymysql
ddt==1.2.1                # via gitpython
doc8==0.8.0
docopt==0.6.2
docutils==0.15
entrypoints==0.3          # via flake8
filelock==3.0.12          # via tox
flake8-isort==2.7.0
flake8==3.7.8             # via flake8-isort
gitdb2==2.0.5             # via gitpython
gitdb==0.6.4              # via gitpython
gitpython==2.1.12         # via bandit
graphene==2.1.7
graphql-core==2.2.1
graphql-relay==2.0.0
graphql-server-core==1.1.1
green==2.16.1
h11==0.8.1
h2==3.1.0                 # via hypercorn
hpack==3.0.0              # via h2
httptools==0.0.13
hypercorn==0.5.4 ; python_version >= "3.6"
hyperframe==5.2.0         # via h2
idna==2.8
imagesize==1.1.0          # via sphinx
immutables==0.9
importlib-metadata==0.18  # via pluggy, pytest, tox
isort==4.3.21             # via flake8-isort, pylint
itsdangerous==1.1.0
jinja2==2.10.1
lazy-object-proxy==1.4.1  # via astroid
lxml==4.3.4               # via green
markupsafe==1.1.1
marshmallow==2.19.5
mccabe==0.6.1             # via flake8, pylint
more-itertools==7.2.0     # via pytest
multidict==4.5.2          # via quart
mypy-extensions==0.4.1    # via mypy
mypy==0.720
nose2==0.9.1
packaging==19.0           # via pytest, sphinx, tox
parse==1.12.0
pbr==5.4.1                # via stevedore
pip-tools==3.9.0
pkginfo==1.5.0.1          # via twine
pluggy==0.12.0            # via pytest, tox
promise==2.2.1
py==1.8.0                 # via pytest, tox
pybadges==2.2.0
pycodestyle==2.5.0        # via autopep8, flake8
pycparser==2.19           # via cffi
pyflakes==2.1.1           # via flake8
pygments==2.4.2
pylint==2.3.1
pymysql==0.9.2            # via aiomysql
pyparsing==2.4.1          # via packaging
pypika==0.30.0
pytest==5.0.1
python-multipart==0.0.5
pytoml==0.1.21            # via hypercorn
pytz==2019.1              # via babel
pyyaml==5.1.1
quart==0.6.13 ; python_version >= "3.6"
readme-renderer==24.0     # via twine
requests-toolbelt==0.9.1
requests==2.22.0
responder==1.3.1
restructuredtext-lint==1.3.0  # via doc8
rfc3986==1.3.2
rope==0.14.0
rx==1.6.1
six==1.12.0
smmap2==2.0.5             # via gitdb2
smmap==0.9.0              # via gitdb
snowballstemmer==1.9.0    # via sphinx
sortedcontainers==2.1.0   # via quart
sphinx-rtd-theme==0.4.3
sphinx==2.1.2
sphinxcontrib-applehelp==1.0.1  # via sphinx
sphinxcontrib-devhelp==1.0.1  # via sphinx
sphinxcontrib-htmlhelp==1.0.2  # via sphinx
sphinxcontrib-httpdomain==1.7.0
sphinxcontrib-jsmath==1.0.1  # via sphinx
sphinxcontrib-qthelp==1.0.2  # via sphinx
sphinxcontrib-serializinghtml==1.1.3  # via sphinx
starlette==0.10.7
stevedore==1.30.1         # via bandit, doc8
testfixtures==6.10.0      # via flake8-isort
toml==0.10.0              # via black, tox
tortoise-orm==0.12.7
tox==3.13.2
tqdm==4.32.2              # via twine
twine==1.13.0
typed-ast==1.4.0          # via astroid, mypy
typesystem==0.2.4
typing-extensions==3.7.4  # via hypercorn, mypy
unidecode==1.1.1          # via green
urllib3==1.25.3
uvicorn==0.8.4
uvloop==0.12.2
virtualenv==16.7.0        # via tox
wcwidth==0.1.7            # via pytest
webencodings==0.5.1       # via bleach
websockets==7.0
werkzeug==0.15.5
whitenoise==4.1.3
wrapt==1.11.2             # via astroid
wsproto==0.14.1           # via hypercorn
zipp==0.5.2               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools==41.0.1        # via sphinx, twine
